Initial typing for infocard component
